Datagenerator and added Test2 project
