+++ GirafRest.Test/Controllers/DepartmentControllerTest.cs
@@ -6,11 +6,11 @@ using System.Collections.Generic;
 using GirafRest.Controllers;
 using GirafRest.Models.Responses;
 using GirafRest.Test.Mocks;
-using static GirafRest.Test.UnitTestExtensions;
+using static GirafRest.Test.DataGenerator;
 using Xunit.Abstractions;
 using System.Linq;
 using GirafRest.Services;
-using static GirafRest.Test.UnitTestExtensions.TestContext;
+using static GirafRest.Test.DataGenerator.TestContext;
 
 namespace GirafRest.Test
 {
