+++ GirafRest.Test/Controllers/PictogramControllerTest.cs
@@ -7,7 +7,7 @@ using GirafRest.Controllers;
 using System;
 using Xunit.Abstractions;
 using GirafRest.Test.Mocks;
-using static GirafRest.Test.UnitTestExtensions;
+using static GirafRest.Test.DataGenerator;
 using GirafRest.Models.DTOs;
 using System.IO;
 using GirafRest.Models.Responses;
