+++ GirafRest.Test/Mocks/MockDbContext.cs
@@ -1,4 +1,5 @@
 using GirafRest.Data;
+using Microsoft.EntityFrameworkCore;
 
 namespace GirafRest.Test.Mocks
 {
