+++ GirafRest.Test/Mocks/MockUserManager.cs
@@ -8,7 +8,7 @@ using GirafRest.Models;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using static GirafRest.Test.UnitTestExtensions;
+using static GirafRest.Test.DataGenerator;
 using System.Linq;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
