+++ GirafRest.Test2/UnitTest1.cs
@@ -1,0 +1,14 @@
+using System;
+using Xunit;
+
+namespace GirafRest.Test2
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
