+++ GirafRest.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GirafRest", "GirafRest\Gira
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GirafRest.Test", "GirafRest.Test\GirafRest.Test.csproj", "{42FEAD29-715B-46BB-A8EE-3594F778CA3C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GirafRest.Test2", "GirafRest.Test2\GirafRest.Test2.csproj", "{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,18 @@ Global
 		{42FEAD29-715B-46BB-A8EE-3594F778CA3C}.Release|x64.Build.0 = Release|Any CPU
 		{42FEAD29-715B-46BB-A8EE-3594F778CA3C}.Release|x86.ActiveCfg = Release|Any CPU
 		{42FEAD29-715B-46BB-A8EE-3594F778CA3C}.Release|x86.Build.0 = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|x64.Build.0 = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Debug|x86.Build.0 = Debug|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|x64.ActiveCfg = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|x64.Build.0 = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|x86.ActiveCfg = Release|Any CPU
+		{76C5AE54-AD3D-442D-8CF3-EF9F4B8516A8}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
